requirement.txt
asgiref==3.7.2e
backports.zoneinfo==0.2.1
Django==4.2.8
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.1.0
dnspython==2.4.2
mongoengine==0.27.0
PyJWT==2.8.0
pymongo==3.12.3
pytz==2023.3.post1
sqlparse==0.2.4
typing-extensions==4.9.0
tzdata==2023.3
